Move e.i.l.c.image.Metadata.encapsulateXMP() to e.i.l.c.image.xmp.Utils
