+++ src/main/java/edu/illinois/library/cantaloupe/processor/codec/jpeg/Util.java
@@ -1,6 +1,6 @@
 package edu.illinois.library.cantaloupe.processor.codec.jpeg;
 
-import edu.illinois.library.cantaloupe.image.Metadata;
+import edu.illinois.library.cantaloupe.image.xmp.Utils;
 import edu.illinois.library.cantaloupe.util.ArrayUtils;
 import edu.illinois.library.cantaloupe.util.StringUtils;
 import org.apache.jena.rdf.model.Model;
@@ -28,7 +28,7 @@ final class Util {
         try {
             final ByteArrayOutputStream os = new ByteArrayOutputStream();
             final byte[] headerBytes = Constants.STANDARD_XMP_SEGMENT_HEADER;
-            final byte[] xmpBytes = Metadata.encapsulateXMP(xmp).
+            final byte[] xmpBytes = Utils.encapsulateXMP(xmp).
                     getBytes(StandardCharsets.UTF_8);
             // write segment marker
             os.write(Marker.APP1.marker());
